add Refactor of Books related functionalities from app.rb
